Move format.ts to utils/ as format-date-time.ts
